1、重构了send模块的方法，交易集合中添加了一个交易conbase交易，2、重构了NewTranaction模块，增加FindSpendOutput寻找所有交易输出/余额方法，改变为寻找未花费的交易输出FindUTXOs、再改变为FindSpendOutputs寻找够本次交易的交易输出+金额  3、增加UTXO模型，可用来存储未消费的交易输出，4、增加了一些小方法，实例化结构体的方法，模仿的是java的有参构造

+++ transaction/txOutPut.go
@@ -12,7 +12,7 @@ import "bytes"
  **/
 type OutPut struct {
 	//比特币的金额
-	Value int64
+	Value uint
 	//锁定脚本（每一笔交易会产生一个锁定脚本，限定于把这笔交易谁能使用）
 	ScriptPubKey []byte
 }
